refactor(shared): move web utils needed by the in-app-agent runtime to shared (LFE-14487)
Mechanical prep for moving the in-app-agent executor into
packages/shared/src/ee (follow-up PR): relocate the MIT-licensed web utilities
the runtime depends on into @langfuse/shared, leaving one-line re-export shims
in web so all existing importers are untouched. Content is copied verbatim;
only import paths change.
