+++ packages/shared/src/features/filters/internalEnvironments.ts
@@ -1,0 +1,14 @@
+import { LangfuseInternalTraceEnvironment } from "../../server/llm/types";
+
+export const DEFAULT_SIDEBAR_HIDDEN_ENVIRONMENTS = [
+  LangfuseInternalTraceEnvironment.PromptExperiments,
+  LangfuseInternalTraceEnvironment.LLMJudge,
+  LangfuseInternalTraceEnvironment.CodeEval,
+  LangfuseInternalTraceEnvironment.NaturalLanguageFilter,
+  "langfuse-evaluation",
+  "sdk-experiment",
+] as const;
+
+export const DEFAULT_SIDEBAR_IMPLICIT_ENVIRONMENT_CONFIG = {
+  hiddenEnvironments: DEFAULT_SIDEBAR_HIDDEN_ENVIRONMENTS,
+} as const;
