+++ packages/shared/src/index.ts
@@ -49,6 +49,17 @@ export { BatchTableNames } from "./interfaces/tableNames";
 // in-app agent
 export * from "./features/inAppAgent/types";
 
+// rbac
+export * from "./features/rbac/projectAccessRights";
+
+// filters
+export * from "./features/filters/internalEnvironments";
+export * from "./features/filters/filterQueryEncoding";
+export * from "./features/filters/sessionPositionInTrace";
+
+// date ranges
+export * from "./utils/dateRanges";
+
 // annotation
 export * from "./features/annotation/types";
