+++ packages/shared/src/server/index.ts
@@ -50,6 +50,9 @@ export * from "./llm/utils";
 export * from "./llm/types";
 export * from "./llm/promptToolConfig";
 export * from "./llm/internalTraceEvents";
+export * from "./llm/langfuseAiCompletion";
+export * from "./utils/baseUrl";
+export * from "./utils/productUrl";
 export * from "./llm/compileChatMessages";
 export * from "./llm/testModelCall";
 export * from "./llm/baseUrlValidation";
