+++ packages/shared/src/utils/stringChecks.ts
@@ -41,3 +41,11 @@ export function stringifyValue(value: unknown) {
       return JSON.stringify(value);
   }
 }
+
+export function truncate(str: string, n = 16) {
+  // '...' suffix if the string is longer than n
+  if (str.length > n) {
+    return str.substring(0, n) + "...";
+  }
+  return str;
+}
