+++ web/src/features/filters/constants/internal-environments.ts
@@ -1,14 +1,4 @@
-import { LangfuseInternalTraceEnvironment } from "@langfuse/shared";
-
-export const DEFAULT_SIDEBAR_HIDDEN_ENVIRONMENTS = [
-  LangfuseInternalTraceEnvironment.PromptExperiments,
-  LangfuseInternalTraceEnvironment.LLMJudge,
-  LangfuseInternalTraceEnvironment.CodeEval,
-  LangfuseInternalTraceEnvironment.NaturalLanguageFilter,
-  "langfuse-evaluation",
-  "sdk-experiment",
-] as const;
-
-export const DEFAULT_SIDEBAR_IMPLICIT_ENVIRONMENT_CONFIG = {
-  hiddenEnvironments: DEFAULT_SIDEBAR_HIDDEN_ENVIRONMENTS,
-} as const;
+export {
+  DEFAULT_SIDEBAR_HIDDEN_ENVIRONMENTS,
+  DEFAULT_SIDEBAR_IMPLICIT_ENVIRONMENT_CONFIG,
+} from "@langfuse/shared";
