+++ web/src/utils/json.ts
@@ -1,27 +1,1 @@
-export function safeJsonParse(value: string): unknown {
-  try {
-    return JSON.parse(value);
-  } catch {
-    return undefined;
-  }
-}
-
-export function stableJsonStringify(value: unknown): string {
-  return JSON.stringify(sortJsonValue(value)) ?? "undefined";
-}
-
-function sortJsonValue(value: unknown): unknown {
-  if (Array.isArray(value)) {
-    return value.map(sortJsonValue);
-  }
-
-  if (value && typeof value === "object") {
-    return Object.fromEntries(
-      Object.entries(value)
-        .sort(([left], [right]) => left.localeCompare(right))
-        .map(([key, nestedValue]) => [key, sortJsonValue(nestedValue)]),
-    );
-  }
-
-  return value;
-}
+export { safeJsonParse, stableJsonStringify } from "@langfuse/shared";
