+++ web/src/utils/string.ts
@@ -1,11 +1,5 @@
+export { truncate } from "@langfuse/shared";
+
 export function lastCharacters(str: string, n: number) {
   return str.substring(str.length - n);
-}
-
-export function truncate(str: string, n = 16) {
-  // '...' suffix if the string is longer than n
-  if (str.length > n) {
-    return str.substring(0, n) + "...";
-  }
-  return str;
 }
